requirements.txt
cryptography==3.4.8
Flask==2.0.1
PyMySQL==1.0.2
PyYAML==5.4.1
